added methods for Bomberman class
- added movementController
- collision detection
- time dependent

+++ src/main/java/com/example/bomberman/gameEngine/AbstractGame.java
@@ -3,7 +3,7 @@ package com.example.bomberman.gameEngine;
 public abstract class AbstractGame {
 
   //TODO:
-  public abstract void update(float deltaTime);
+  public abstract void update(double deltaTime);
 
 //  public abstract void render(GameContainer gc, Renderer renderer);
 }
